Fixed big with createModel function and null treeItem.
createModelFunction was always returning treeItem->data(row) as the
parent, but the root treeItem is set to nullptr and therefore has no
function or data to access.  This caused a segfault in Windows.  The
function now checks of tree Item is nullptr and returns nullptr if it
is, and the data pertaining to the row if not.